+++ app/components/generic-rdfa-variable/insert-menu.js
@@ -1,0 +1,1 @@
+export { default } from '@lblod/ember-rdfa-editor-lblod-plugins/components/generic-rdfa-variable/insert-menu';