+++ addon/plugins/variable-plugin/utils/constants.ts
@@ -56,7 +56,7 @@ export const DEFAULT_VARIABLE_TYPES: Record<string, VariableType> = {
   },
   date: {
     label: 'date',
-    constructor: (schema, label: 'date') => {
+    constructor: (schema, label = 'date') => {
       return unwrap(
         schema.nodes.date.createAndFill({
           mappingResource: `http://data.lblod.info/mappings/${uuidv4()}`,