+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add label to variables
 - Number variable - support min/max
+### Dependencies
+- Bumps `webpack` from 5.75.0 to 5.76.0
 
 ## [7.1.0] - 2023-05-18