+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Number variable - support min/max
 ### Dependencies
 - Bumps `webpack` from 5.75.0 to 5.76.0
+- Move `tracked-toolbox` to dependencies
 
 ## [7.1.0] - 2023-05-18